refactor: move common functions to util.go.
