+++ cmd/argo/commands/clustertemplate/get.go
@@ -1,17 +1,11 @@
 package clustertemplate
 
 import (
-	"encoding/json"
-	"fmt"
 	"log"
-
-	"github.com/argoproj/pkg/humanize"
-	"github.com/spf13/cobra"
-	"sigs.k8s.io/yaml"
 
 	"github.com/argoproj/argo-workflows/v3/cmd/argo/commands/client"
 	clusterworkflowtmplpkg "github.com/argoproj/argo-workflows/v3/pkg/apiclient/clusterworkflowtemplate"
-	wfv1 "github.com/argoproj/argo-workflows/v3/pkg/apis/workflow/v1alpha1"
+	"github.com/spf13/cobra"
 )
 
 func NewGetCommand() *cobra.Command {
@@ -40,27 +34,4 @@ func NewGetCommand() *cobra.Command {
 
 	command.Flags().StringVarP(&output, "output", "o", "", "Output format. One of: json|yaml|wide")
 	return command
-}
-
-func printClusterWorkflowTemplate(wf *wfv1.ClusterWorkflowTemplate, outFmt string) {
-	switch outFmt {
-	case "name":
-		fmt.Println(wf.ObjectMeta.Name)
-	case "json":
-		outBytes, _ := json.MarshalIndent(wf, "", "    ")
-		fmt.Println(string(outBytes))
-	case "yaml":
-		outBytes, _ := yaml.Marshal(wf)
-		fmt.Print(string(outBytes))
-	case "wide", "":
-		printClusterWorkflowTemplateHelper(wf)
-	default:
-		log.Fatalf("Unknown output format: %s", outFmt)
-	}
-}
-
-func printClusterWorkflowTemplateHelper(wf *wfv1.ClusterWorkflowTemplate) {
-	const fmtStr = "%-20s %v\n"
-	fmt.Printf(fmtStr, "Name:", wf.ObjectMeta.Name)
-	fmt.Printf(fmtStr, "Created:", humanize.Timestamp(wf.ObjectMeta.CreationTimestamp.Time))
 }
