+++ cmd/argo/commands/template/util.go
@@ -1,10 +1,17 @@
 package template
 
 import (
+	"encoding/json"
+	"fmt"
 	"log"
 
+	argoJson "github.com/argoproj/pkg/json"
+	"sigs.k8s.io/yaml"
+
 	wfv1 "github.com/argoproj/argo-workflows/v3/pkg/apis/workflow/v1alpha1"
+	"github.com/argoproj/argo-workflows/v3/workflow/common"
 	"github.com/argoproj/argo-workflows/v3/workflow/util"
+	"github.com/argoproj/pkg/humanize"
 )
 
 func generateWorkflowTemplates(filePaths []string, strict bool) []wfv1.WorkflowTemplate {
@@ -24,4 +31,47 @@ func generateWorkflowTemplates(filePaths []string, strict bool) []wfv1.WorkflowT
 	}
 
 	return workflowTemplates
+}
+
+// unmarshalWorkflowTemplates unmarshals the input bytes as either json or yaml
+func unmarshalWorkflowTemplates(wfBytes []byte, strict bool) []wfv1.WorkflowTemplate {
+	var wf wfv1.WorkflowTemplate
+	var jsonOpts []argoJson.JSONOpt
+	if strict {
+		jsonOpts = append(jsonOpts, argoJson.DisallowUnknownFields)
+	}
+	err := argoJson.Unmarshal(wfBytes, &wf, jsonOpts...)
+	if err == nil {
+		return []wfv1.WorkflowTemplate{wf}
+	}
+	yamlWfs, err := common.SplitWorkflowTemplateYAMLFile(wfBytes, strict)
+	if err == nil {
+		return yamlWfs
+	}
+	log.Fatalf("Failed to parse workflow template: %v", err)
+	return nil
+}
+
+func printWorkflowTemplate(wf *wfv1.WorkflowTemplate, outFmt string) {
+	switch outFmt {
+	case "name":
+		fmt.Println(wf.ObjectMeta.Name)
+	case "json":
+		outBytes, _ := json.MarshalIndent(wf, "", "    ")
+		fmt.Println(string(outBytes))
+	case "yaml":
+		outBytes, _ := yaml.Marshal(wf)
+		fmt.Print(string(outBytes))
+	case "wide", "":
+		printWorkflowTemplateHelper(wf)
+	default:
+		log.Fatalf("Unknown output format: %s", outFmt)
+	}
+}
+
+func printWorkflowTemplateHelper(wf *wfv1.WorkflowTemplate) {
+	const fmtStr = "%-20s %v\n"
+	fmt.Printf(fmtStr, "Name:", wf.ObjectMeta.Name)
+	fmt.Printf(fmtStr, "Namespace:", wf.ObjectMeta.Namespace)
+	fmt.Printf(fmtStr, "Created:", humanize.Timestamp(wf.ObjectMeta.CreationTimestamp.Time))
 }
